backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
pydantic==2.7.1
python-multipart==0.0.9
httpx==0.27.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1